Refactor package structure for improved organization
- Renamed and reorganized classes into 'slack' and 'config' packages to enhance clarity.
- Updated package declarations to reflect the new paths.
- Imported necessary classes in existing files to align with the refactored structure.
