+++ src/main/java/eu/kratochvil/jidlobot/SlackBotService.java
@@ -7,8 +7,12 @@ import com.slack.api.methods.SlackApiException;
 import com.slack.api.methods.request.chat.ChatPostMessageRequest;
 import com.slack.api.methods.response.chat.ChatPostMessageResponse;
 import com.slack.api.model.event.AppMentionEvent;
+import eu.kratochvil.jidlobot.config.ApplicationConfig;
+import eu.kratochvil.jidlobot.config.SlackConfig;
 import eu.kratochvil.jidlobot.model.DailyMenu;
 import eu.kratochvil.jidlobot.parser.DailyMenuParser;
+import eu.kratochvil.jidlobot.slack.BotMessageBuilder;
+import eu.kratochvil.jidlobot.slack.DailyMenuMessageBuilder;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import org.jetbrains.annotations.NotNull;
