+++ src/main/java/eu/kratochvil/jidlobot/config/SlackConfig.java
@@ -1,4 +1,4 @@
-package eu.kratochvil.jidlobot;
+package eu.kratochvil.jidlobot.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
