+++ src/main/java/eu/kratochvil/jidlobot/slack/BotMessageBuilder.java
@@ -1,4 +1,4 @@
-package eu.kratochvil.jidlobot;
+package eu.kratochvil.jidlobot.slack;
 
 import com.slack.api.methods.request.chat.ChatPostMessageRequest;
 import com.slack.api.model.block.Blocks;
